A simple UI in which the user can adjust parameters of the 2-alpha probabilistic model to see how changes in parameter values affect the shape of the model's curves
When the program starts up, the user will be prompted to enter the leaf number and upper bound for the days post inoculation axis; since
the data collection stops at 10, it is recommended to use a value of 11 or 12 for best visibility, but to
see the log-term behavior of the models a large value can be used. When the graph first appears, the model
curves will not show. The user must click on one of the sliders first, upon which the model curves will be
drawn. After this, the user can slide the parameter values around to their heart's content.

 - The first panel shows the model curves and the experimental data.
 - The second panel shows the nll value for the model using the current set of parameters. Note that though only
 one leaf at a time can be shown, the total nll depends on all four, even the three not on screen
 - The third panel shows coinfection plotted against Venus-only, both the models and experimental data. The
 best-fit linear approximation of the data is shown in red, along with grey reference lines. The 'm' values
 show the slopes of their respective lines.
 - The fourth panel shows the slope of the tail of the model curve in the third panel.
 - The fifth panel shows coinfection plotted against BFP-only, both the models and experimental data. The
 best-fit linear approximation of the data is shown in red, along with grey reference lines. The 'm' values
 show the slopes of their respective lines.
 - The sixth panel shows the slope of the tail of the model curve in the fifth panel.